Process the i386 and amd64 versions of the Debian and Ubuntu packages separately
This change is required because - depending on the package name -
dump_syms might find the wrong debug information for a given library.
For example it might try to dump the symbols for a 32-bit library using
the 64-bit debug information and vice-versa.

This changes causes the packages to be processed on a per-architecture
basis which solves the issue.