Fixes two issues when dealing when downloading multiple drivers
First of all it fixes the maximum number of drivers that get downloaded
in one go. Since we updated the counter in a subshell we'd download up
to 10 drivers per vendor, while we want to only download a maximum of 10
drivers per pass.

The second issue is that we did not delete the index.html file between
different vendors, this would lead us to only ever scrape AMD drivers.