Refactor the scripts to only attempt to fetch packages that we haven't processed yet
We remove the step where we generate fake files and filter the URLs
we're fetching instead. This has several advantages: the fake package
generation step was slow and didn't scale, as we'd pile up more and more
files in SHA256SUMS over time. Fetching the packages themselves is also
faster, as we don't even query for packages that we know we've already
downloaded. Last but not least the resulting SHA256SUMS file shrinks to
only contain the packages that exist at any given moment.