Correctly process different versions of Alpine Linux packages with the same name
Alpine Linux packages under different release channels might have the
same name if they correspond to the same version of an upstream package.
This changes takes this into account and processes the various channels
as if they were different distros.

This change also adds more packages to the list and fixes an issue with
package names containing the '+' character not being processed
correctly.